Matches no longer contain an offset.
Calculating the offset every time a match is created can slow down the
parse dramatically. Also, the actual offset of a match is not usually
very useful during interpretation.

The functionality was moved into citrus/debug.rb to preserve the ability
to inspect the offset of a match when debugging.