Added Match#captures.
Match#captures is the preferred method of retrieving named sub-captures
when performing match evaluation. It's easier to use than Match#matches
because it provides direct access to matches as they were written in the
rule definition, without requiring the user to think about how the
actual match tree was constructed.

This commit deprecates Match#name, Match#names, Match#has_name?,
Match#find, and match naming in general. Matches should not keep track
of the names of the rules that created them since that information can
be derived as needed from the event stream as demonstrated by
Match#captures.

Also, Citrus::NoMatchError was eliminated to favor a more fluid style of
use when using Match#method_missing to find a capture of a given name.