Try to consume the entire input by default
Instead of defaulting to false, the consume option now defaults to true.
This is the behavior that many users expect the first time they use
Citrus because it is the way that Treetop works. It used to be expensive
to calculate the length of a Match, so we defaulted to false. It is now
trivial, so it shouldn't slow things down to default to true.