Added Rule#elide?
This restores the former behavior to matches that are the results of a
Choice rule, which is that they are essentially the same match as is
returned by the first matching rule of that choice. This largely
eliminates the need to call Match#first in the value blocks of choices,
which feels more natural when writing them.