Added Citrus.require
Added the ability for require statements inside Citrus files to require
other Citrus files that may be in the $LOAD_PATH. Also, modified the
behavior of Citrus.load to cache the result of already loaded files so
that we don't load the same file twice.

Modified examples/ipaddress.citrus to demonstrate the use of require
from within Citrus files.