Adds self to Match#captures
A common point of confusion for some Citrus users is that they are not
able to retrieve a match by calling its own name on itself. This commit
adds self to the internal @captures Hash in the first position.

Also, Match#to_a and Match#[] were added for convenience when retrieving
captures and submatches.