Keep a master offset in Match instead of slicing.
Instead of successively slicing strings in matches, we keep the
parsed text under @source, and an offset in the later in @offset.

Btw, that @offset might be very useful for keeping trace of the
localisation of the match in the source text (e.g. for subsequent
semantic passes; semantic error detection for instance).