Show why Match#method_missing is unsafe with a test.
The added test case shows that Kernel::p is actually called in place
of Match#method_missing when 'p' is used as a label or rule name.

As any gem might contribute private methods to Object, using Citrus
labels (or even rule names, actually) in semantic productions without
doing so with Match#captures is unsafe.

Moreover, the addition of any method/helper to the Match class may
potentially break an existing grammar, so it breaks Citrus's public
API and should probably increment the MAJOR version number.